requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
babel==2.17.0
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
blis==1.0.1
bokeh==3.6.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.1.0
cmake==4.0.3
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
confection==0.1.5
contourpy==1.2.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.8
dask==2024.11.2
dask-expr==1.1.19
datasets==3.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distributed==2024.11.2
distro==1.9.0
dlib @ file:///C:/Users/mayou/Downloads/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
docutils==0.21.2
et_xmlfile==2.0.0
Euler-Deconvolution @ file:///D:/Documents/Semister%206/mueed%20task/Euler_Deconvolution_project
executing==2.2.0
experta==1.9.4
face-recognition==1.3.0
face-recognition-models==0.3.0
Faker==28.0.0
fastjsonschema==2.21.1
filelock==3.16.1
fire==0.6.0
Flask==3.0.3
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.66.1
gTTS==2.5.4
h11==0.14.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
hyperframe==5.2.0
id==1.5.0
idna==2.10
imbalanced-learn==0.12.3
importlib_metadata==8.5.0
imutils==0.5.4
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.2
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
keras-attention==1.0.0
Keras-Preprocessing==1.1.2
keyring==25.6.0
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
llvmlite==0.44.0
locket==1.0.0
lxml==5.2.1
lz4==4.3.3
magnetic-transform @ file:///D:/Documents/Semister%206/mueed%20task/Grav_Mag_Transform_project
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.2
ml-dtypes==0.4.0
more-itertools==10.6.0
mpi4py==4.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.8
narwhals==1.8.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neattext==0.1.3
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.21
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
openai==1.47.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pdf2docx==0.5.8
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
polars==1.33.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyahocorasick==2.1.0
pyarrow==17.0.0
pycaw==20240210
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyopencl==2025.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-engineio==4.12.3
python-json-logger==3.2.1
python-socketio==5.14.2
pytools==2025.1.2
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
schema==0.6.7
scikit-learn==1.6.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.38.0
sympy==1.13.3
tblib==3.0.0
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
textsearch==0.0.24
tf_keras==2.17.0
Theano==1.0.5
thinc==8.3.2
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
twine==6.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.3.193
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3==2.2.1
virtualenv==20.29.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.1
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.9.4
zict==3.0.0
zipp==3.21.0
Flask
gunicorn
vercel-python
opencv-python
face-recognition
numpy
imutils
